requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.3
django-heroku==0.3.1
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
sqlparse==0.5.2
typing_extensions==4.12.2
whitenoise==6.8.2
python-dotenv==1.0.1
django-filter==24.3